This version adds a brand new lex/yacc parser to decode test cases. It is implemented in tbparser.py
Also, a major change was introduced in systools. The new version, now
keeps track of all processes sharing the same process group id or
their descendants.

Finally, with the aim of improving generality of the testbot, the
snippet for checking whether a valid solution was generated or not has
been fully removed. Validating solutions should be a different process