Until now, the dbparser recognized two different types of tables: sys and data. The first contained data that is gathered at every ping of the underlying executable (such as cpu/wall clock time, memory consumed, number of processes and threads); the latter was filled with data obtained once the executable was over with every test case. Yet, BotTestCase implements a third type of data: admin datatables which hold administrative information.
In this changeset a new type of tables has been implemented: user
datatables which are filled by the client code of autobot. The
mechanism is simple: user datatables shall be prefixed with 'user_' in
the database specification file and the user has to add the rows (in
the form of tables) to the dictionary 'stats' with the key that
matches the name of the user datatable.