Finally, I could work on the notion of namespaces. There are still some work to do with the user namespace (see notes below) but the main load of work is finally done.
The modifications performed in this changeset close a part of Issue
#11: Database Parser that should be split into different issues to
maintain information about the other ideas originally exposed there.

Now, there are four different namespaces fully operational: namespace,
data namespace, user namespace and param namespace. They are created
and maintained internally by autobot.

On the other hand, there are different types of variables recognized
by the database parser:

* sysvar: these are variables computed by autobot at every cycle of the
          execution of the solver. Every 'check' seconds the new contents
          of these variables is computed

* mainvar: these are the flags given to the main script using autobot
           (ie., testbot) These variables can be used to create a template
           for the 'output' file

* datavar: data processed from the stdout of the executable. These data is
           retrieved using different regular expressions and they are
           processed only once after the execution of the solver over
           every test case

* filevar: these variables are given as filenames whose value are the
           contents of the file

* param: these are the flags given to the executable. They are retrieved
         from the specification of the current test case under execution

* dirvar: these are also the flags given to the executable but they are
          named after their position

To make this distinction more apparent, the database parser recognizes
now also prefixes in addition to the one-character shortcuts that were
already implemented:

 type of variable   prefix
 -----------------+-----------
 sysvar           | 'sys.'
 datavar          | 'data.'
 dirvar           | 'dir.'
 filevar          | 'file.'
 mainvar          | "main.'
 param            | 'param.'
 -----------------+-----------

Thus, this changeset explicitly maintains a relationship between
variable types and namespaces described below:

 namespace   preffixes
 ----------+-----------------
 namespace | sysvar mainvar
 data      | datavar filevar
 user      | --
 param     | param, dirvar
 ----------+-----------------

Because data is now split into various namespaces, the __call__ method
of the Prologue/Epilogue and the Enter/Windup classes do also receive
these namespaces only in case they are useful (in some cases, it does
not make sense to distribute that information):

 * Enter: it receives namespace and user namespace

 * Prologue: it receives the namespace, the user namespace and the
             param namespace

 * Epilogue: it receives all namespaces: the main namespace, the data
             namespace, the user namespace and the param namespace

 * WindUp: it receives the namespace, the data namespace and the user
           namespace

With the aim of showing how namespaces can be handled by the user,
these classes have been programmed now also in testbot. They just
simply print out additional information (with the contents of these
namespaces and other useful information) in case the debug level
information is requested.

Note that the user namespace is distributed always and autobot never
use it (to avoid clashing with names chosen by the user). However, it
is not recognized by the database parser yet so that it is not allowed
to request writing user data into the databases (this was originally
done just by creating "user" datatables but this does not make any
sense at all and following Alvaro's advice it makes more sense to
consider this data ordinary data).