Finally, regexp are accepted by the db specification language. They can be inserted anywhere in db specification files and used in the specification of columns of database tables ---it is not mandatory at all for the regexps to be defined before they are used.
Regexps are defined with the format:

regexp <name> <specification-string>

and columns can use regexps with the format:

<name>.<group>

where <group> are groups defined in the <specification-string> with
the directive (?P<group>...). These columns are qualified with a
vartype 'REGEXP'.

The default regexp used to process generic output is still in use and
it should be allowed to overwrite its specification. This generic
regexp is given the name 'data' and therefore it can be overwritten
with the statement:

regexp data <specification-string>

Of course, regexps can accept an arbitrary number of <groups>.

Right now, however, regexps are not used in any particular way. They
are correctly parsed but this version does not allow database table
specifications to refer to regexps since they are not processed yet.