Finally, autobot accepts and processes regexps as defined in the database specification file. Everytime a line from the standard output matches a regexp, its contents are written into the 'regexp' namespace which, in turn, is passed in the Epilogue and WindUp functions.
Matches of regexps are written in the 'regexp' namespace as multi-key
attributes which contain a list with all matches of the same regexp
(i.e., everything is preserved!) whose keynames are exactly the same
than the group names appearing in the regexp definition.

Namespace was modified also to allow setattr/getattr methods to handle
keys with length 1 in multi-key attributes.