This changeset contains the skeleton for the parsebot.
The parsebot extends the definition of an experiment to consider also
automatically parsing text files as if they were the output generated
by a different program. This eases collaborative work. In this new
setting the definitions used are:

* Experiment: an experiment denotes a particular solver (among an
              arbitrary large collection of them) over its tests cases
              or the parsing process of all text files.

* Run: denotes a particular execution of a solver over a given test
       case or the parsing process of a single text file.

The match between executing solvers or automatically parsing files
applies to the whole design of autobot. Therefore, parsebot also
implements the same control flow implemented in testbot, ie.,
third-party code is allowed to use the services provided by the
testbot:

* By invoking BotMain. As a result, specific parsebots can be
  implemented with their own setUp/test_* methods/tearDown

* By providing specific prologue/epilogue enter/windUp definitions

Additionally, this changeset contains a basic definition of an
argument parser for parsebot