This is the first release that fully supports regular expressions with contexts. A few preliminary tests have been performed where the first context is another regular expression or it is an ordinary variable (whose value is the retrieved from a namespace)
Also, from the previous version, text files are parsed more
efficiently. Instead of matching all declared regular expressions with
the contents of the given file, the data tables are processed and
regular expressions are matched agains the contents of the current
file *only in case they appear in the definition of a
column*. Additionally, regular expressions are checked first for
existence. In case they have been already processed, they are not
processed again.