I have significantly enhanced the programming of snippets. From a general point of view, snippets are a structure recognized by the bots on their own:
1. Snippets are now defined in the dbparser to have the same methods
than regexps so that they both can be dealt with the same way

2. The definition of snippets recognizes now two keywords in the
output variables: 'volatile' and 'static'. All output variables are
static by default. If an output variable is declared volatile, the
snippet shall be evaluated again when autobot is writing to the
database.

This facility is not implemented yet however.

3. The botparser contains now a snippet namespace for storing its own
results

4. The poll method in dbparser receives now the snippet namespace also
so that it can access its variables

5. dbexpression has now a dedicated method to eval snippets

These modifications provide more flexibility to the definition and
usage of snippets