Still working on snippets and regexps. The previous version was extremely complex (so much that it was not even feasible to perform some simple operations).
This release is the first move towards an easier implementation of
regexps and snippets.

The main change is that regexps are now defined as a collection of
contexts whose head can be anything but a regexp. All items but the
head are regexps that should have been previously defined. From now
one, if a regexp refers to the context currently processed, it should
be prefixed by sys.stdout.

Thus, the main task of botparser is to ensure that all variables
referenced in all columns of all datatables and also the heads of all
regexps are correctly processed. The latter involves snippets and
filevars. But regexps are never processed by botparser.

As a consequence, a lot of code were deleted (and most likely more
code will be deleted in a near future).

I also completed the definition of start/end parse/fullparse times