Volatile and static output variables have been completely coded.
The flow is as follows: botparser is in charge of ensuring that all
necessary data is present in the namespaces to evaluate any
dbexpression. In case a snippet is found (either because a column is a
snippet on its own or it appears as the head of a regexp), then it is
not evaluated if the snippet is volatile (ie., at least one of its
output variables is declared as volatile). Otherwise, it is evaluated
in case it has not been evaluated yet. If a snippet is not evaluated
because it was volatile, then poll will request its evaluation prior
to the evaluation of the column.

Experiments with the current time using 'time.sleep (1)' in the python
snippet prove that volatile variables change their value in every
invocation.