First version of condorize.py.
This script takes the same arguments than testbot.py and it creates a
condor submission description file that automatically adds it to the
condor queue. It ignores, however, the following flags:
`--show-placeholders`, `--parse-tests` and `--parse-db`

Still, it does not copy the testbot installation is required
(`--copy-files`). Finally, it has not been tested yet on a real condor
pool.