Additional modifications to README.md
Also, it contains simple modifications to bottester.py that enables
the execution of testbot.py with very simple dbspecs (ie., those
containing only sys tables). The reason why other more complex db
tables are not allowed is because the dbparser was lately modified to
accommodate changes necessary for parsebot which have to be included
in testbot now ---mainly by making testbot to use the services
provided by parsebot.